feat(node): add crypto KeyObject support
