+++ types/node/tsconfig.json
@@ -51,7 +51,8 @@
         "test/net.ts",
         "test/dgram.ts",
         "test/buffer.ts",
-        "test/tty.ts"
+        "test/tty.ts",
+        "test/crypto.ts"
     ],
     "compilerOptions": {
         "module": "commonjs",
